dev/art-e-stepwise/pyproject.toml
[project]
name = "art-e"
version = "0.1.0"
description = "Deep learning research for email search and processing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.4.1",
    "diskcache>=5.6.3",
    "huggingface-hub>=0.29.3",
    "ipykernel>=6.29.5",
    "kaggle>=1.5.12",
    "langchain-core>=0.3.51",
    "mail-parser>=3.15.0",
    "matplotlib>=3.10.1",
    "pandas>=1.3.0",
    "panza>=0.1.0",
    "pip>=25.0.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.62.0",
    "transformers>=4.50.3",
    "skypilot[runpod]==0.9.3",
    "ipywidgets>=8.1.6",
    "polars>=1.27.1",
    "openpipe-art[backend,plotting]",
    "weave>=0.51.53",
    "accelerate==1.7.0",
    "litellm==1.73.6",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["art_e*"]

[tool.uv]
override-dependencies = [
    "bitsandbytes; sys_platform == 'linux'",
    "xformers; sys_platform == 'linux'",
    "vllm; sys_platform == 'linux'",
]


[tool.uv.sources]
openpipe-art = { path = "../../", editable = true }

[dependency-groups]
dev = ["art-e", "ruff>=0.12.3"]
